variadic_function-1

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -22,7 +22,7 @@ times = n;
 sum = 0;
 while (times--)
 {
-int cur_num = va_arg(nums, size_t);
+int cur_num = va_arg(nums, int);
 sum += cur_num;
 }
 va_end(nums);
